rename test_harness to test_util
